Optionally support other Wire objects.
Many processors, now days have more than one hardware I2C on them.  Sometimes it is convienent to use a different I2C object like Wire1 or Wire2 instead of the default Wire object.

This change allows you to pass in a pointer to the desired I2C(Wire) object on the constructor. which defaults to Wire.

This change requires the main header file to include the wire.h header file.